+++ lib/src/carvables/analyzer.dart
@@ -1,10 +1,11 @@
 import 'package:analyzer/dart/ast/ast.dart';
+import 'package:analyzer/dart/ast/syntactic_entity.dart';
 import 'package:analyzer/dart/element/element.dart';
 import 'package:carvable/src/utils/ast.dart';
 import 'package:carvable/src/utils/iterable.dart';
 
 import './string.dart';
-import '../carvings/analyzer/node.dart';
+import '../carvings/analyzer/token.dart';
 import '../carvings/analyzer/element.dart';
 import '../carvings/string/replace.dart';
 
@@ -14,8 +15,8 @@ abstract class CarvableAnalyzer extends CarvableString {
 
 	/* -= Modifying Methods =- */
 
-  /// Remove an AstNode from the resulting source.
-	CarvableAnalyzer removeNode(AstNode node) => this..carve(node.carving);
+  /// Remove a token from the resulting source.
+	CarvableAnalyzer removeToken(SyntacticEntity token) => this..carve(token.carving);
 
   /// Remove an Element from the resulting source.
 	CarvableAnalyzer removeElement(Element element) {
@@ -24,8 +25,8 @@ abstract class CarvableAnalyzer extends CarvableString {
 		return this;
 	}
 
-  /// Remove multiple AstNodes from the resulting source.
-	CarvableAnalyzer removeNodes(Iterable<AstNode> nodes) => this..carveAll(nodes.map((node) => node.carving));
+  /// Remove multiple tokens from the resulting source.
+	CarvableAnalyzer removeTokens(Iterable<SyntacticEntity> tokens) => this..carveAll(tokens.map((token) => token.carving));
   
   /// Remove multiple Elements from the resulting source.
 	CarvableAnalyzer removeElements(Iterable<Element> elements) =>
@@ -34,17 +35,17 @@ abstract class CarvableAnalyzer extends CarvableString {
 		);
 }
 
-class CarvableNode extends CarvableAnalyzer {
+class CarvableToken extends CarvableAnalyzer {
 	final AstNode target;
-	CarvableNode(this.target) : super.empty();
+	CarvableToken(this.target) : super.empty();
 
 	@override String get input => target.toSource();
 
-  @override CarvableNode removeNode(AstNode node) => super.removeNode(node) as CarvableNode;
-  @override CarvableNode removeNodes(Iterable<AstNode> nodes) => super.removeNodes(nodes) as CarvableNode;
+  @override CarvableToken removeToken(SyntacticEntity token) => super.removeToken(token) as CarvableToken;
+  @override CarvableToken removeTokens(Iterable<SyntacticEntity> tokens) => super.removeTokens(tokens) as CarvableToken;
 
-  @override CarvableNode removeElement(Element element) => super.removeElement(element) as CarvableNode;
-  @override CarvableNode removeElements(Iterable<Element> elements) => super.removeElements(elements) as CarvableNode;
+  @override CarvableToken removeElement(Element element) => super.removeElement(element) as CarvableToken;
+  @override CarvableToken removeElements(Iterable<Element> elements) => super.removeElements(elements) as CarvableToken;
 }
 
 class CarvableElement extends CarvableAnalyzer {
@@ -53,8 +54,8 @@ class CarvableElement extends CarvableAnalyzer {
 
 	@override String get input => target.source?.contents.data ?? '';
 
-  @override CarvableElement removeNode(AstNode node) => super.removeNode(node) as CarvableElement;
-  @override CarvableElement removeNodes(Iterable<AstNode> nodes) => super.removeNodes(nodes) as CarvableElement;
+  @override CarvableElement removeToken(SyntacticEntity token) => super.removeToken(token) as CarvableElement;
+  @override CarvableElement removeTokens(Iterable<SyntacticEntity> tokens) => super.removeTokens(tokens) as CarvableElement;
 
   @override CarvableElement removeElement(Element element) => super.removeElement(element) as CarvableElement;
   @override CarvableElement removeElements(Iterable<Element> elements) => super.removeElements(elements) as CarvableElement;
@@ -62,26 +63,29 @@ class CarvableElement extends CarvableAnalyzer {
 
 /* -= Extensions =- */
 
+extension CarvingTokenExtension on SyntacticEntity {
+  CarvingToken get carving => CarvingToken(this);
+}
+
 extension CarvableNodeExtension on AstNode {
-	CarvableAnalyzer get carvable => CarvableNode(this);
-	CarvingNode get carving => CarvingNode(this);
+	CarvableAnalyzer get carvable => CarvableToken(this);
 
-	/// Remove an AstNode from the resulting source.
-	CarvableAnalyzer remove(AstNode node) => removeNode(node);
+	/// Remove a token from the resulting source.
+	CarvableAnalyzer remove(SyntacticEntity token) => removeToken(token);
 
-	/// Remove an AstNode from the resulting source.
-	CarvableAnalyzer removeNode(AstNode node) => carvable.removeNode(node);
+	/// Remove a token from the resulting source.
+	CarvableAnalyzer removeToken(SyntacticEntity token) => carvable.removeToken(token);
 
 	/// Remove an Element from the resulting source.
 	CarvableAnalyzer removeElement(Element element) =>
 		carvable.removeElement(element);
 
-	/// Remove multiple AstNodes from the resulting source.
-	CarvableAnalyzer removeAll(Iterable<AstNode> node) => removeAllNodes(node);
+	/// Remove multiple tokens from the resulting source.
+	CarvableAnalyzer removeAll(Iterable<SyntacticEntity> token) => removeAllNodes(token);
 
-	/// Remove multiple AstNodes from the resulting source.
-	CarvableAnalyzer removeAllNodes(Iterable<AstNode> nodes) =>
-		carvable.removeNodes(nodes);
+	/// Remove multiple tokens from the resulting source.
+	CarvableAnalyzer removeAllNodes(Iterable<SyntacticEntity> tokens) =>
+		carvable.removeTokens(tokens);
 
 	/// Remove multiple Elements from the resulting source.
 	CarvableAnalyzer removeAllElements(Iterable<Element> elements) =>
@@ -92,20 +96,20 @@ extension CarvableElementExtension on Element {
 	CarvableElement get carvable => CarvableElement(this);
 	CarvingElement get carving => CarvingElement(this);
 
-	/// Remove an AstNode from the resulting source.
-	CarvableAnalyzer remove(AstNode node) => removeNode(node);
+	/// Remove a token from the resulting source.
+	CarvableAnalyzer remove(SyntacticEntity token) => removeTokens(token);
 
-	/// Remove an AstNode from the resulting source.
-	CarvableAnalyzer removeNode(AstNode node) => carvable.removeNode(node);
+	/// Remove a token from the resulting source.
+	CarvableAnalyzer removeTokens(SyntacticEntity token) => carvable.removeToken(token);
 
 	/// Remove an Element from the resulting source.
 	CarvableAnalyzer removeElement(Element element) => carvable.removeElement(element);
 
-	/// Remove multiple AstNodes from the resulting source.
-	CarvableAnalyzer removeAll(Iterable<AstNode> node) => removeAllNodes(node);
+	/// Remove multiple tokens from the resulting source.
+	CarvableAnalyzer removeAll(Iterable<SyntacticEntity> token) => removeAllTokens(token);
 
-	/// Remove multiple AstNodes from the resulting source.
-	CarvableAnalyzer removeAllNodes(Iterable<AstNode> nodes) => carvable.removeNodes(nodes);
+	/// Remove multiple tokens from the resulting source.
+	CarvableAnalyzer removeAllTokens(Iterable<SyntacticEntity> tokens) => carvable.removeTokens(tokens);
 
 	/// Remove multiple Elements from the resulting source.
 	CarvableAnalyzer removeAllElements(Iterable<Element> elements) => carvable.removeElements(elements);
@@ -115,20 +119,20 @@ extension CarvableLibraryExtension on LibraryElement {
 	CarvableAnalyzer get carvable => CarvableElement(this);
 	CarvingReplacement get carving => CarvingReplacement(0, source.contents.data.length);
 
-	/// Remove an AstNode from the resulting source.
+	/// Remove an Element from the resulting source.
 	CarvableAnalyzer remove(Element element) => removeElement(element);
 
-	/// Remove an AstNode from the resulting source.
-	CarvableAnalyzer removeNode(AstNode node) => carvable.removeNode(node);
+	/// Remove a token from the resulting source.
+	CarvableAnalyzer removeToken(SyntacticEntity token) => carvable.removeToken(token);
 
 	/// Remove an Element from the resulting source.
 	CarvableAnalyzer removeElement(Element element) => carvable.removeElement(element);
 
-	/// Remove multiple AstNodes from the resulting source.
+	/// Remove multiple Elements from the resulting source.
 	CarvableAnalyzer removeAll(Iterable<Element> elements) => removeAllElements(elements);
 
-	/// Remove multiple AstNodes from the resulting source.
-	CarvableAnalyzer removeAllNodes(Iterable<AstNode> nodes) => carvable.removeNodes(nodes);
+	/// Remove multiple tokens from the resulting source.
+	CarvableAnalyzer removeAllNodes(Iterable<SyntacticEntity> tokens) => carvable.removeTokens(tokens);
 
 	/// Remove multiple Elements from the resulting source.
 	CarvableAnalyzer removeAllElements(Iterable<Element> elements) => carvable.removeElements(elements);